Break consensus_inner from consensus and fix a anticone computation bug
